requirements.txt
PyYAML==3.11
pytest==2.8.3
